brand_predictor/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless
Pillow
gradio
ultralytics
scipy
numpy
tqdm
